Visualize results for patch selection and delete pyc files
