+++ mtsmorf/move_exp/functions/move_experiment_functions.py
@@ -28,7 +28,8 @@ def _preprocess_epochs(epochs, resample_rate=500):
     new_epochs = epochs.filter(l_freq=1, h_freq=fs / 2 - 1)
 
     # Downsample epochs to 500 Hz
-    new_epochs = new_epochs.resample(resample_rate)
+    if resample_rate is None and isinstance(resample_rate, (int, float)):
+        new_epochs = new_epochs.resample(resample_rate)
 
     return new_epochs
